+++ file.h
@@ -4,12 +4,12 @@
 /*estrutura que será preenchidas com os dados do arquivo*/
 typedef struct VALUES{
     int n;//Números de pontos tabelados para x e y
-    float* x;
-    float* y;
+    double* x;
+    double* y;
     int a;//intervalos da integral a e b
     int b;
     int i;//números de pontos a sofrerem interpolação p
-    float* p;
+    double* p;
     int t;//quantidade de trapezios
 }values;
 /*função para ler o arquivo e retornar um vetor com os valores necessários*/
